+++ Sources/HTTP/Parser/ParseResults.swift
@@ -1,23 +1,36 @@
 import URI
 import CHTTP
 
+/// Possible header states
 enum HeaderState {
     case none
     case value(key: HeaderKey, Bytes)
     case key(Bytes)
 }
 
+/// The parse results object helps get around
+/// the issue of not being able to capture context
+/// with C closures. 
+///
+/// All C closures must be sent some object that 
+/// this parse results object can be retreived from.
+/// 
+/// See the convenience methods below to see how the
+/// object is set and fetched from the C object.
 internal final class ParseResults {
+    // state
+    var headerState: HeaderState
     var isComplete: Bool
-    var url: Bytes
+    
+    // message components
+    var version: Version?
     var headers: [HeaderKey: String]
     var body: Bytes
+
+    // url
+    var url: Bytes
     
-    var uri: URI?
-    var version: Version?
-    
-    var headerState: HeaderState
-    
+    /// Creates a new results object
     init() {
         self.isComplete = false
         self.url = []
@@ -25,29 +38,22 @@ internal final class ParseResults {
         self.body = []
         self.headerState = .none
     }
+}
+
+// MARK: Convenience
+
+extension ParseResults {
+    /// Sets the parse results object on a C parser
+    static func set(_ results: inout ParseResults, on parser: inout http_parser) {
+        parser.data = UnsafeMutableRawPointer(&results)
+    }
     
-    static func from(_ parser: UnsafePointer<http_parser>?) -> ParseResults? {
+    /// Fetches the parse results object from the C parser
+    static func get(from parser: UnsafePointer<http_parser>?) -> ParseResults? {
         return parser?
             .pointee
             .data
             .assumingMemoryBound(to: ParseResults.self)
             .pointee
-    }
-    
-    func assertResults() throws -> (
-        version: Version,
-        uri: URI,
-        headers: [HeaderKey: String],
-        body: Body
-    ) {
-        guard let version = self.version else {
-            throw ParserError.noVersion
-        }
-        
-        guard let uri = self.uri else {
-            throw ParserError.noURI
-        }
-        
-        return (version, uri, headers, .data(body))
     }
 }
