+++ Sources/HTTP/Parser/ResponseParser.swift
@@ -2,31 +2,34 @@ import Transport
 import CHTTP
 import URI
 
-public final class ResponseParser<Stream: DuplexStream>: Parser {
+/// Parses responses from a readable stream.
+public final class ResponseParser<Stream: ReadableStream>: CHTTPParser {
+    // Internal variables to conform
+    // to the C HTTP parser protocol.
     typealias StreamType = Stream
     let stream: Stream
     var parser: http_parser
     var settings: http_parser_settings
     var buffer: Bytes
     
+    /// Creates a new Response parser.
     public init(stream: Stream) {
         self.stream = stream
         self.parser = http_parser()
         self.settings = http_parser_settings()
         http_parser_init(&parser, HTTP_RESPONSE)
         self.buffer = Bytes()
-        self.buffer.reserveCapacity(2048)
+        self.buffer.reserveCapacity(bufferSize)
     }
     
-    
+    /// Parses a Response from the stream.
     public func parse() throws -> Response {
-        http_parser_init(&parser, HTTP_RESPONSE)
         let results = try parseMessage()
         
         let status = Status(statusCode: Int(parser.status_code))
         
         guard let version = results.version else {
-            throw ParserError.noVersion
+            throw ParserError.invalidMessage
         }
         
         let response = Response(
@@ -37,5 +40,19 @@ public final class ResponseParser<Stream: DuplexStream>: Parser {
         )
         
         return response
+    }
+}
+
+// MARK: Settings
+
+private var _bufferSize = 2048
+extension ResponseParser {
+    public var bufferSize: Int {
+        get {
+            return _bufferSize
+        }
+        set {
+            _bufferSize = newValue
+        }
     }
 }
